+++ MinhaAPI/Domain/Models/CarrinhoModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int QtdProduto { get; set; }
-        public double ValorTotalProduto { get; set; }
+        public decimal ValorTotalProduto { get; set; }
 
         public int ProdutoId { get; set; }
         public ProdutoModel Produto { get; set; }
